# Improve the heuristic on picking distributed size.