As #8980 described, this PR changes the heuristic on the distributed size to take the vector size into account. When the vector size is provided, it tries to find `the largest multiplier of the vector size which can divide the dim size and <= dim size / 2`.

Since tiling with a multiplier of vector size doesn't always have benefits, the vector size is passed as an optional hint per dimension. This PR only enables the new heuristic for the Conv2d on output feature dimension.

On x86 with AVX512:
|                                   | Base(ms) | After(ms) | Speedup |
|-----------------------------------|----------|-----------|---------|
| MobileNetV3, dylib-sync, 1 thread | 6.13     | 5.12      | 1.19x   |
| MobileNetV3, dylib, 4 threads     | 4.27     | 3.94      | 1.08x   |

On the sample kernel `main_dispatch_33` from MobileNetV3:
```mlir
func @main_dispatch_33(%arg0: tensor<1x18x18x240xf32> {linalg.buffer_layout = affine_map<(d0, d1, d2, d3) -> (d0, d1, d2, d3)>, linalg.inplaceable = false}, %arg1: tensor<5x5x240xf32> {linalg.buffer_layout = affine_map<(d0, d1, d2, d3) -> (d0, d1, d2, d3)>, linalg.inplaceable = false}, %arg2: tensor<1x14x14x240xf32> {linalg.buffer_layout = affine_map<(d0, d1, d2, d3) -> (d0, d1, d2, d3)>, linalg.inplaceable = true}) -> tensor<1x14x14x240xf32> {
  %0 = linalg.depthwise_conv_2d_nhwc_hwc {dilations = dense<1> : tensor<2xi64>, strides = dense<1> : tensor<2xi64>} ins(%arg0, %arg1 : tensor<1x18x18x240xf32>, tensor<5x5x240xf32>) outs(%arg2 : tensor<1x14x14x240xf32>) -> tensor<1x14x14x240xf32>
  return %0 : tensor<1x14x14x240xf32>
}
```
|                                        | Base(ms) | After(ms) | Speedup |
|----------------------------------------|----------|-----------|---------|
| main_dispatch_33, dylib-sync, 1 thread | 27.2     | 2.43      | 11.19x  |
| main_dispatch_33, dylib, 4 threads     | 10.5     | 3.07      | 3.42x   |

(Compiled with `-iree-hal-benchmark-dispatch-repeat-count=100`)